Move evaluator to toolpad-core
