+++ packages/toolpad-core/src/runtime.tsx
@@ -1,7 +1,15 @@
 import * as React from 'react';
 import ErrorIcon from '@mui/icons-material/Error';
 import { RUNTIME_PROP_NODE_ID, RUNTIME_PROP_SLOTS } from './constants.js';
-import type { SlotType, ComponentConfig, RuntimeEvent } from './index';
+import type {
+  SlotType,
+  ComponentConfig,
+  RuntimeEvent,
+  ArgTypeDefinition,
+  BindableAttrValue,
+  LiveBinding,
+  RuntimeError,
+} from './types';
 
 declare global {
   interface Window {
@@ -46,11 +54,6 @@ function PlaceholderWrapper(props: PlaceholderWrapperProps) {
       }}
     />
   );
-}
-
-export interface RuntimeError {
-  message: string;
-  stack?: string;
 }
 
 export interface NodeRuntimeWrapperProps {
@@ -225,4 +228,44 @@ export async function importCodeComponent<P>(
   // eslint-disable-next-line no-underscore-dangle
   (Component as any).__config = config;
   return Component;
+}
+
+let iframe: HTMLIFrameElement;
+function evalCode(code: string, globalScope: Record<string, unknown>) {
+  if (!iframe) {
+    iframe = document.createElement('iframe');
+    iframe.setAttribute('sandbox', 'allow-same-origin allow-scripts');
+    iframe.style.display = 'none';
+    document.documentElement.appendChild(iframe);
+  }
+
+  // eslint-disable-next-line no-underscore-dangle
+  (iframe.contentWindow as any).__SCOPE = globalScope;
+  return (iframe.contentWindow as any).eval(`with (window.__SCOPE) { ${code} }`);
+}
+
+export function evaluateBindable<V>(
+  bindable: BindableAttrValue<V> | null,
+  globalScope: Record<string, unknown>,
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  argType?: ArgTypeDefinition,
+): LiveBinding {
+  const execExpression = () => {
+    if (bindable?.type === 'jsExpression') {
+      return evalCode(bindable?.value, globalScope);
+    }
+
+    if (bindable?.type === 'const') {
+      return bindable?.value;
+    }
+
+    return undefined;
+  };
+
+  try {
+    return { value: execExpression() };
+  } catch (err) {
+    console.error(`Oh no`, err);
+    return { error: err as Error };
+  }
 }
